ref: separate Mappers' tests from API tests and rename DTO to RequestDTO

+++ src/main/java/com/veteroch4k/tasktracker/models/TimeRecord.java
@@ -8,9 +8,9 @@ public class TimeRecord {
 
   private Long id;
 
-  private int employeeId;
+  private Long employeeId;
 
-  private int taskId;
+  private Long taskId;
 
   private LocalDateTime startTime;
